// src/generated/models/Cr871_vendorsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr871_vendorscr871_currentriskband = {
  144610000: 'LowRisk',
  144610001: 'MediumRisk',
  144610002: 'HighRisk',
  144610003: 'CriticalRisk'
} as const;
export type Cr871_vendorscr871_currentriskband = keyof typeof Cr871_vendorscr871_currentriskband;
export const Cr871_vendorscr871_currentriskrating = {
  144610000: 'Critical',
  144610001: 'High',
  144610002: 'Medium',
  144610003: 'Low'
} as const;
export type Cr871_vendorscr871_currentriskrating = keyof typeof Cr871_vendorscr871_currentriskrating;
export const Cr871_vendorscr871_vendorstatus = {
  144610000: 'Active',
  144610001: 'Inactive',
  144610002: 'Offboarded'
} as const;
export type Cr871_vendorscr871_vendorstatus = keyof typeof Cr871_vendorscr871_vendorstatus;
export const Cr871_vendorsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr871_vendorsstatecode = keyof typeof Cr871_vendorsstatecode;
export const Cr871_vendorsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr871_vendorsstatuscode = keyof typeof Cr871_vendorsstatuscode;

export interface Cr871_vendorsBase {
  cr871_assessmentcount?: number;
  cr871_currentoverallscore?: number;
  cr871_currentriskband?: Cr871_vendorscr871_currentriskband;
  cr871_currentriskrating?: Cr871_vendorscr871_currentriskrating;
  cr871_lastassesseddate?: string;
  cr871_legalentityname?: string;
  cr871_nextassessmentdue?: string;
  cr871_onboardeddate?: string;
  cr871_primaryassessoremail?: string;
  cr871_rcnumber?: string;
  cr871_vendorcontactemail?: string;
  cr871_vendorid: string;
  cr871_vendorname?: string;
  cr871_vendornotes?: string;
  cr871_vendorstatus?: Cr871_vendorscr871_vendorstatus;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr871_vendorsstatecode;
  statuscode?: Cr871_vendorsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr871_vendors extends Cr871_vendorsBase {
  cr871_currentriskbandname?: string;
  cr871_currentriskratingname?: string;
  cr871_vendorstatusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}